Move the checks to become the action framework
* The plan is to move each check to a separate file within the actions directory.
* Note that the post-conversion checks are staying in checks.py for now.
  We may move them to the actions framework or we might not.  This needs
  to be evaluated in a later release.
* Migrate check_convert2rhel_latest.
  Start migrating check_convert2rhel_latest to the actions framework to see
  what we need to make this work.
